pyproject.toml
[project]
name = "laxyfile"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.7.0",
    "keyboard>=0.13.5",
    "openai>=1.0.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "psutil>=5.9.0",
    "pydantic>=2.4.0",
    "pygments>=2.16.0",
    "python-magic>=0.4.27",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "rich>=13.7.0",
    "send2trash>=1.8.2",
    "textual>=0.41.0",
    "typer>=0.9.0",
    "watchdog>=3.0.0",
]
